requirements.txt
# 百速一键AI视频生成 - 依赖列表

# Flask及扩展
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.2.3

# 数据库
SQLAlchemy==2.0.4

# 视频处理
moviepy==1.0.3
Pillow==9.4.0

# AI服务
zhipuai==1.0.7
dashscope==1.10.0

# 环境变量
python-dotenv==1.0.0

# 工具库
requests==2.28.2
numpy==1.24.2